fix model

+++ Entities/ViewModels/Products/ProductMongoDbModel.cs
@@ -57,7 +57,7 @@ namespace Entities.ViewModels.Products
         public string description_ingredients { get; set; }
         public string description_effect { get; set; }
         public string description_usepolicy { get; set; }
-        public int? review_count { get; set; }
+        public double? review_count { get; set; }
         public float? rating { get; set; }
         public long? total_sold { get; set; }
